+++ meute.js
@@ -1,0 +1,67 @@
+meute = (function() {
+  var masterPwd, sockethubClient, config = {}, configDone = {};
+
+  function connectFurther() {
+    if (config.sockethub) {
+      if (configDone.sockethub) {
+        for (i in config) {
+          if (!configDone[i]) {
+            sockethubClient.set(i, config[i]);
+            configDone[i] = true;
+          }
+        }
+      } else {
+        sockethubClient = SockethubClient.connect(config);
+        sockethubClient.on('registered', function() {
+          configDone.sockethub = true;
+          connectFurther();
+        });
+      }
+    } else {
+      //nothing to do without a sockethub config
+    }
+  }
+ 
+  function bootstrap() {
+    var modulesToTry = {
+      sockethub: true,
+      irc: true,
+      twitter: true,
+      facebook: true,
+      email: true
+    };
+    if (remoteStorage) {
+      remoteStorage.on('ready', function() {
+        for (i in modulesToTry) {
+          loadAccount(i);
+        }
+      });
+    }
+  }
+  function loadAccount(which) {
+    remoteStorage[which].getConfig(masterPwd).then(function(config) {
+      addAccount(which, config);
+    }
+  }
+   
+  function setMasterPassword(pwd) {
+    masterPwd = pwd;
+    bootstrap();
+  }
+  function addAccount(which, config) {
+    config[which] = config;
+    connectFurther();
+    if (remoteStorage[which] && remoteStorage[which].setConfig) {
+      remoteStorage[which].setConfig(masterPwd, config);
+    }
+  }
+  function send(obj) {
+    sockethubClient.send(obj);
+  }
+  
+  return {
+    setMasterPassword: setMasterPassword,
+    addAccount: addAccount,
+    send: send
+  };
+})();