+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+
+* first version of the meute.setMasterPassword function
+* first version of the meute.addAccount function
+* first version of the meute.send function
+
 # 0.1.0
 
 * a loose and eclectic collection of scripts, very unorganized and not suitable for reuse in other apps